algorithm/recommend_system/s02_recsys_neucf/pyproject.toml
[tool.poetry]
name = "generate_recommendation"
version = "0.1.0"
description = "Generate Recommendation and Cache"
readme = "README.md"
authors = ["shuntakochi <shuntakochi@gmail.com>"]
packages = [
    { include = "database"},
    { include = "recommender"},
    { include = "utils"},
    { include = "neucf"}
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
ipykernel = "^6.29.5"
sqlalchemy = "^2.0.37"
polars = "^1.19.0"
s3fs = "^2024.12.0"
boto3 = "<1.36.0"
matplotlib = "^3.10.0"
psycopg2-binary = "^2.9.10"
japanize-matplotlib = "^1.1.3"
black = "^24.10.0"
isort = "^6.0.0"
mypy = "^1.14.1"
haversine = "^2.9.0"
requests = "^2.32.3"
filelock = "3.18.0"
fsspec = "2024.12.0"
Jinja2 = "3.1.6"
MarkupSafe = "3.0.2"
mpmath = "1.3.0"
numpy = "2.2.3"
packaging = "24.2"
pandas = "2.2.3"
python-dateutil = "2.9.0.post0"
tqdm = "4.67.1"
supabase = "^2.15.0"
dotenv = "^0.9.9"
uvicorn = "^0.34.0"
fastapi = "^0.115.12"
blobfile = "^3.0.0"
asyncpg = "^0.30.0"
mangum = "^0.19.0"
redis = "^6.0.0"
valkey-glide = "^1.3.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
lint.select = ["F401", "F403", "F404"]  # 未使用importや存在しないimportなど
